refactor: 디비 초기화 클래스 패키지 이동
